+++ src/codeql.ts
@@ -23,6 +23,7 @@ import {
   getTrapCachingExtractorConfigArgsForLang,
 } from "./trap-caching";
 import * as util from "./util";
+import { wrapError } from "./util";
 
 type Options = Array<string | number | boolean>;
 
@@ -288,12 +289,6 @@ export const CODEQL_VERSION_SECURITY_EXPERIMENTAL_SUITE = "2.12.1";
 export const CODEQL_VERSION_INIT_WITH_QLCONFIG = "2.12.4";
 
 /**
- * Versions 2.12.6+ of the CodeQL CLI fix a bug where duplicate notification objects could be produced,
- * leading to an invalid SARIF output.
- */
-export const CODEQL_VERSION_DUPLICATE_NOTIFICATIONS_FIXED = "2.12.6";
-
-/**
  * Set up CodeQL CLI access.
  *
  * @param toolsInput
@@ -345,7 +340,7 @@ export async function setupCodeQL(
       toolsVersion,
     };
   } catch (e) {
-    logger.error(e instanceof Error ? e : new Error(String(e)));
+    logger.error(wrapError(e).message);
     throw new Error("Unable to download and extract CodeQL CLI");
   }
 }
@@ -762,12 +757,8 @@ export async function getCodeQLForCmd(
         Feature.ExportDiagnosticsEnabled,
         this
       );
-      const shouldWorkaroundInvalidNotifications =
-        shouldExportDiagnostics &&
-        !(await util.codeQlVersionAbove(
-          this,
-          CODEQL_VERSION_DUPLICATE_NOTIFICATIONS_FIXED
-        ));
+      // Update this to take into account the CodeQL version when we have a version with the fix.
+      const shouldWorkaroundInvalidNotifications = shouldExportDiagnostics;
       const codeqlOutputFile = shouldWorkaroundInvalidNotifications
         ? path.join(config.tempDir, "codeql-intermediate-results.sarif")
         : sarifFile;
@@ -917,11 +908,8 @@ export async function getCodeQLForCmd(
       tempDir: string,
       logger: Logger
     ): Promise<void> {
-      const shouldWorkaroundInvalidNotifications =
-        !(await util.codeQlVersionAbove(
-          this,
-          CODEQL_VERSION_DUPLICATE_NOTIFICATIONS_FIXED
-        ));
+      // Update this to take into account the CodeQL version when we have a version with the fix.
+      const shouldWorkaroundInvalidNotifications = true;
       const codeqlOutputFile = shouldWorkaroundInvalidNotifications
         ? path.join(tempDir, "codeql-intermediate-results.sarif")
         : sarifFile;